refactor(agent): restructure agent charm code to use modules for non event based functions
