+++ agent/charms/testflinger-agent-host-charm/src/charm.py
@@ -7,16 +7,13 @@
 
 import logging
 import os
-import re
 import shutil
-import socket
 import sys
-from base64 import b64decode
 from pathlib import Path
-from typing import Optional, Tuple
 
+import supervisord
 import testflinger_source
-from common import run_with_logged_errors
+from common import copy_ssh_keys, run_with_logged_errors, update_charm_scripts
 from defaults import (
     AGENT_CONFIGS_PATH,
     LOCAL_TESTFLINGER_PATH,
@@ -34,9 +31,6 @@ from charms.operator_libs_linux.v0 import apt, passwd
 logger = logging.getLogger(__name__)
 
 SUPERVISOR_CONF_DIR = "/etc/supervisor/conf.d"
-TESTFLINGER_AGENT_CMD = "testflinger-agent"
-DEFAULT_PORT_RANGE_START = 8000
-DEFAULT_PORT_RANGE_END = 10000
 
 
 class TestflingerAgentHostCharm(CharmBase):
@@ -142,146 +136,6 @@ class TestflingerAgentHostCharm(CharmBase):
             shutil.rmtree(repo_path, ignore_errors=True)
         shutil.move(tmp_repo_path, repo_path)
 
-    def get_supervisor_agents_port_mapping(self) -> dict[str, int]:
-        """Get current agent:port mapping from supervisor config files.
-
-        Returns:
-            dict: Mapping of agent names to their configured ports.
-                 Empty dict if no valid configurations found.
-        """
-        agent_port_mapping = {}
-        supervisor_conf_dir = Path(SUPERVISOR_CONF_DIR)
-
-        if not supervisor_conf_dir.exists():
-            logger.debug(
-                "Supervisor config directory %s does not exist",
-                supervisor_conf_dir,
-            )
-            return agent_port_mapping
-
-        try:
-            for conf_file in supervisor_conf_dir.iterdir():
-                if conf_file.suffix == ".conf":
-                    conf_path = str(conf_file)
-                    try:
-                        agent_name, port = self.parse_supervisor_config_file(
-                            conf_path
-                        )
-                        if agent_name and port:
-                            agent_port_mapping[agent_name] = port
-                            logger.debug(
-                                "Found configured agent %s on port %s",
-                                agent_name,
-                                port,
-                            )
-                    except (OSError, IOError) as e:
-                        logger.debug(
-                            "Failed to read config file %s: %s", conf_path, e
-                        )
-                        continue
-                    except ValueError as e:
-                        logger.debug(
-                            "Failed to parse port number in %s: %s",
-                            conf_path,
-                            e,
-                        )
-                        continue
-        except (OSError, IOError) as e:
-            logger.warning("Failed to read supervisor config files: %s", e)
-
-        return agent_port_mapping
-
-    def parse_supervisor_config_file(
-        self, conf_path: str
-    ) -> Tuple[Optional[str], Optional[int]]:
-        """Extract agent name and port from supervisor config file.
-
-        Returns:
-            tuple: (agent_name, port) if both found, (None, None) otherwise.
-        """
-        try:
-            agent_name = None
-            port = None
-
-            with open(conf_path, "r", encoding="utf-8") as f:
-                for raw_line in f:
-                    line = raw_line.strip()
-
-                    if line.startswith("[program:"):
-                        program_match = re.search(
-                            r"\[program:([^\]]+)\]", line
-                        )
-                        if program_match:
-                            agent_name = program_match.group(1)
-                    elif (
-                        line.startswith("command=")
-                        and TESTFLINGER_AGENT_CMD in line
-                    ):
-                        port_match = re.search(r"-p (\d+)", line)
-                        if port_match:
-                            port = int(port_match.group(1))
-
-                    if agent_name and port:
-                        return agent_name, port
-            return None, None
-
-        except (OSError, IOError) as e:
-            logger.debug("Failed to read config file '%s': %s", conf_path, e)
-            return None, None
-        except ValueError as e:
-            logger.debug(
-                "Failed to parse port number in '%s': %s", conf_path, e
-            )
-            return None, None
-
-    def find_available_port(
-        self,
-        used_ports: set[int],
-        start_port: int = DEFAULT_PORT_RANGE_START,
-        max_port: int = DEFAULT_PORT_RANGE_END,
-    ) -> int:
-        """Find an available port avoiding used ports."""
-        for port in range(start_port, max_port):
-            if port in used_ports:
-                continue
-
-            try:
-                with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-                    s.bind(("localhost", port))
-                    return port
-            except OSError:
-                continue
-        raise RuntimeError(
-            f"No available ports found in range {start_port}-{max_port}"
-        )
-
-    def assign_metrics_port(
-        self,
-        agent_name: str,
-        configured_agents: dict[str, int],
-        used_ports: set[int],
-    ) -> int:
-        """Assign a metrics port using supervisor config file discovery."""
-        # Check if agent is already configured with a port
-        if agent_name in configured_agents:
-            current_port = configured_agents[agent_name]
-            logger.info(
-                "Agent %s keeping existing port %s",
-                agent_name,
-                current_port,
-            )
-            return current_port
-        # If not configured, find a new port
-        try:
-            port = self.find_available_port(used_ports)
-            logger.info("Assigned new port %s to agent %s", port, agent_name)
-            return port
-        except RuntimeError as e:
-            logger.error(
-                "Failed to find available port for %s: %s", agent_name, e
-            )
-            raise
-
     def write_supervisor_service_files(self, initial_metrics_port=8000):
         """
         Generate supervisord service files for all agents.
@@ -312,7 +166,7 @@ class TestflingerAgentHostCharm(CharmBase):
             sys.exit(1)
 
         # Get current agents and their ports from the service files
-        configured_agents = self.get_supervisor_agents_port_mapping()
+        configured_agents = supervisord.get_supervisor_agents_port_mapping()
         used_ports = set(configured_agents.values())
         logger.info(
             "Found %d configured agents",
@@ -336,7 +190,7 @@ class TestflingerAgentHostCharm(CharmBase):
             agent_name = agent_dir.name
 
             try:
-                metric_endpoint_port = self.assign_metrics_port(
+                metric_endpoint_port = supervisord.assign_metrics_port(
                     agent_name, configured_agents, used_ports
                 )
                 used_ports.add(metric_endpoint_port)
@@ -368,16 +222,6 @@ class TestflingerAgentHostCharm(CharmBase):
             ) as agent_file:
                 agent_file.write(rendered)
 
-    def supervisor_update(self):
-        """
-        Once supervisord service files have been written, new agents will be
-        automatically started, and missing agents will be removed by running
-        `supervisorctl update`. This only applies to supervisor conf files
-        that have changed. So any agents for which the conf file has not
-        changed will be unaffected.
-        """
-        run_with_logged_errors(["supervisorctl", "update"])
-
     def restart_agents(self):
         """
         Mark all agents as needing a restart when they are not running a job
@@ -390,53 +234,10 @@ class TestflingerAgentHostCharm(CharmBase):
         passwd.add_group("docker")
         passwd.add_user_to_group("ubuntu", "docker")
 
-    def write_file(self, location, contents):
-        with open(location, "w", encoding="utf-8", errors="ignore") as out:
-            out.write(contents)
-
-    def copy_ssh_keys(self):
-        try:
-            ssh_config = self.config.get("ssh-config")
-            self.write_file("/home/ubuntu/.ssh/config", ssh_config)
-            os.chown("/home/ubuntu/.ssh/config", 1000, 1000)
-            os.chmod("/home/ubuntu/.ssh/config", 0o640)
-
-            priv_key = self.config.get("ssh-private-key", "")
-            self.write_file(
-                "/home/ubuntu/.ssh/id_rsa", b64decode(priv_key).decode()
-            )
-            os.chown("/home/ubuntu/.ssh/id_rsa", 1000, 1000)
-            os.chmod("/home/ubuntu/.ssh/id_rsa", 0o600)
-
-            pub_key = self.config.get("ssh-public-key", "")
-            self.write_file(
-                "/home/ubuntu/.ssh/id_rsa.pub", b64decode(pub_key).decode()
-            )
-            os.chown("/home/ubuntu/.ssh/id_rsa.pub", 1000, 1000)
-        except (TypeError, UnicodeDecodeError):
-            logger.error(
-                "Failed to decode ssh keys - ensure they are base64 encoded"
-            )
-            raise
-
     def update_tf_cmd_scripts(self):
         """Update tf-cmd-scripts."""
         self.unit.status = MaintenanceStatus("Installing tf-cmd-scripts")
-        tf_cmd_dir = "src/tf-cmd-scripts/"
-        usr_local_bin = Path("/usr/local/bin")
-
-        for tf_cmd_file in os.listdir(tf_cmd_dir):
-            template = Template(
-                open(os.path.join(tf_cmd_dir, tf_cmd_file)).read()
-            )
-            rendered = template.render(
-                agent_configs_path=AGENT_CONFIGS_PATH,
-                config_dir=self.config.get("config-dir"),
-                virtual_env_path=VIRTUAL_ENV_PATH,
-            )
-            agent_file = usr_local_bin / tf_cmd_file
-            agent_file.write_text(rendered)
-            agent_file.chmod(0o775)
+        update_charm_scripts(self.config)
 
     def on_upgrade_charm(self, _):
         """Upgrade hook."""
@@ -459,10 +260,10 @@ class TestflingerAgentHostCharm(CharmBase):
                 "config-repo and config-dir must be set"
             )
             return
-        self.copy_ssh_keys()
+        copy_ssh_keys(self.config)
         self.update_tf_cmd_scripts()
         self.write_supervisor_service_files()
-        self.supervisor_update()
+        supervisord.supervisor_update()
         self.restart_agents()
         self.unit.status = ActiveStatus()
 
@@ -501,7 +302,7 @@ class TestflingerAgentHostCharm(CharmBase):
             )
             return
         self.write_supervisor_service_files()
-        self.supervisor_update()
+        supervisord.supervisor_update()
         self.restart_agents()
         self.unit.status = ActiveStatus()
